Colosseum API Reference

Documentation for Colosseum.

Colosseum.armDisarm

Colosseum.confirmConnection

Colosseum.getHomeGeoPoint

Colosseum.getMinRequiredClientVersion

Colosseum.isApiControlEnabled

Colosseum.ping

Colosseum.reset

Colosseum.simAddDetectionFilterMeshName

Colosseum.simClearDetectionMeshNames

Colosseum.simContinueForFrames

Colosseum.simContinueForTime

Colosseum.simDestroyObject

Colosseum.simEnableWeather

Colosseum.simGetCameraInfo

Colosseum.simGetCollisionInfo

Colosseum.simGetDetections

Colosseum.simGetDistortionParams

Colosseum.simGetImage

Colosseum.simGetImages

Colosseum.simGetMeshPositionVertexBuffers

Colosseum.simGetObjectPose

Colosseum.simGetObjectScale

Colosseum.simGetSegmentationObjectID

Colosseum.simGetVehiclePose

Colosseum.simGetWorldExtents

Colosseum.simIsPause

Colosseum.simListAssets

Colosseum.simListSceneObjects

Colosseum.simLoadLevel

Colosseum.simPause

Colosseum.simPrintLogMessage

Colosseum.simRunConsoleCommand

Colosseum.simSetCameraFov

Colosseum.simSetCameraPose

Colosseum.simSetDetectionFilterRadius

Colosseum.simSetDistortionParam

Colosseum.simSetDistortionParams

Colosseum.simSetLightIntensity

Colosseum.simSetObjectMaterial

Colosseum.simSetObjectMaterialFromTexture

Colosseum.simSetObjectPose

Colosseum.simSetObjectScale

Colosseum.simSetSegmentationObjectID

Colosseum.simSetTimeOfDay

Colosseum.simSetTraceLine

Colosseum.simSetVehiclePose

Colosseum.simSetWeatherParameter

Colosseum.simSpawnObject

Colosseum.simSwapTextures

Colosseum.simTestLineOfSightBetweenPoints

Colosseum.simTestLineOfSightToPoint

Colosseum.armDisarm — Function

Arms || disarms vehicle

Args: arm (bool) True to arm, false to disarm the vehicle vehicle_name (str, optional) Name of the vehicle to send this command to

Returns: bool: Success

Colosseum.confirmConnection — Method

Checks state of connection every 1 sec and reports it in Console so user can see the progress for connection.

Colosseum.getHomeGeoPoint — Function

Get the Home location of the vehicle

Args: vehicle_name (str, optional) Name of vehicle to get home location of

Returns: GeoPoint: Home location of the vehicle

Colosseum.getMinRequiredClientVersion — Method

Enables || disables API control for vehicle corresponding to vehicle_name

Args: isenabled (bool) True to enable, false to disable API control vehiclename (str, optional) Name of the vehicle to send this command to

Colosseum.isApiControlEnabled — Function

Returns true if API control is established.

If false (which is functionault) then API calls would be ignored. After a successful call to enableApiControl, isApiControlEnabled should return true.

Args: vehicle_name (str, optional) Name of the vehicle

Returns: bool: If API control is enabled

Colosseum.ping — Method

ping(c::VehicleClient)

If connection is established then this call will return true otherwise it will be blocked until timeout

Colosseum.reset — Method

reset(c::VehicleClient)

Reset the vehicle to its original starting state

Note that you must call enableApiControl and armDisarm again after the call to reset

Colosseum.simAddDetectionFilterMeshName — Function

Add mesh name to detect in wild card format

For example: simAddDetectionFilterMeshName("Car*") will detect all instance named "Car*"

Args: cameraname (str) Name of the camera, for backwards compatibility, ID numbers such as 0,1,etc. can also be used imagetype (ImageType) Type of image required meshname (str) mesh name in wild card format vehiclename (str, optional) Vehicle which the camera is associated with external (bool, optional) Whether the camera is an External Camera

Colosseum.simClearDetectionMeshNames — Function

Clear all mesh names from detection filter

Args: cameraname (str) Name of the camera, for backwards compatibility, ID numbers such as 0,1,etc. can also be used imagetype (ImageType) Type of image required vehicle_name (str, optional) Vehicle which the camera is associated with external (bool, optional) Whether the camera is an External Camera

Colosseum.simContinueForFrames — Method

Continue (or resume if paused) the simulation for the specified number of frames, after which the simulation will be paused.

Args: frames (int) Frames to run the simulation for

Colosseum.simContinueForTime — Method

Continue the simulation for the specified number of seconds

Args: seconds (float) Time to run the simulation for

Colosseum.simDestroyObject — Method

Removes selected object from the world

Args: object_name (str) Name of object to be removed

Returns: bool: True if object is queued up for removal

Colosseum.simEnableWeather — Method

Enable Weather effects. Needs to be called before using simSetWeatherParameter API

Args: enable (bool) True to enable, false to disable

Colosseum.simGetCameraInfo — Function

Get details about the camera

Args: cameraname (str) Name of the camera, for backwards compatibility, ID numbers such as 0,1,etc. can also be used vehiclename (str, optional) Vehicle which the camera is associated with external (bool, optional) Whether the camera is an External Camera

Returns: CameraInfo:

Colosseum.simGetCollisionInfo — Function

Args: vehicle_name (str, optional) Name of the Vehicle to get the info of

Returns: CollisionInfo:

Colosseum.simGetDetections — Function

Get current detections

Args: cameraname (str) Name of the camera, for backwards compatibility, ID numbers such as 0,1,etc. can also be used imagetype (ImageType) Type of image required vehicle_name (str, optional) Vehicle which the camera is associated with external (bool, optional) Whether the camera is an External Camera

Returns: DetectionInfo array

Colosseum.simGetDistortionParams — Function

Get camera distortion parameters

Args: cameraname (str) Name of the camera, for backwards compatibility, ID numbers such as 0,1,etc. can also be used vehiclename (str, optional) Vehicle which the camera is associated with external (bool, optional) Whether the camera is an External Camera

Returns: List (float) List of distortion parameter values corresponding to K1, K2, K3, P1, P2 respectively.

Colosseum.simGetImage — Function

Get a single image Returns bytes of png format image which can be dumped into abinary file to create .png image string_to_uint8_array() can be used to convert into Numpy unit8 array See https://microsoft.github.io/AirSim/imageapis/ for details Args: cameraname (str) Name of the camera, for backwards compatibility, ID numbers such as 0,1,etc. can also be used imagetype (ImageType) Type of image required vehiclename (str, optional) Name of the vehicle with the camera external (bool, optional) Whether the camera is an External Camera Returns: Binary string literal of compressed png image

Colosseum.simGetImages — Function

Get multiple images

See https://microsoft.github.io/AirSim/image_apis/ for details and examples

Args: requests (list[ImageRequest]) Images required vehicle_name (str, optional) Name of vehicle associated with the camera external (bool, optional) Whether the camera is an External Camera

Returns: list[ImageResponse]:

Colosseum.simGetMeshPositionVertexBuffers — Method

Returns the static meshes that make up the scene

See https://microsoft.github.io/AirSim/meshes/ for details and how to use this

Returns: list[MeshPositionVertexBuffersResponse]:

Colosseum.simGetObjectPose — Method

The position inside the returned Pose is in the world frame

Args: object_name (str) Object to get the Pose of

Returns: Pose:

Colosseum.simGetObjectScale — Method

Gets scale of an object in the world

Args: object_name (str) Object to get the scale of

Returns: airsim.Vector3r: Scale

Colosseum.simGetSegmentationObjectID — Method

Returns Object ID for the given mesh name

Mapping of Object IDs to RGB values can be seen at https://microsoft.github.io/AirSim/seg_rgbs.txt

Args: mesh_name (str) Name of the mesh to get the ID of

Colosseum.simGetVehiclePose — Function

The position inside the returned Pose is in the frame of the vehicle's starting point

Args: vehicle_name (str, optional) Name of the vehicle to get the Pose of

Returns: Pose:

Colosseum.simGetWorldExtents — Method

Returns a list of GeoPoints representing the minimum and maximum extents of the world

Returns: list[GeoPoint]

Colosseum.simIsPause — Method

Returns true if the simulation is paused

Returns: bool: If the simulation is paused

Colosseum.simListAssets — Method

Lists all the assets present in the Asset Registry

Returns: list[str]: Names of all the assets

Colosseum.simListSceneObjects — Function

Lists the objects present in the environment

end

functionault behaviour is to list all objects, regex can be used to return smaller list of matching objects || actors

Args:
    name_regex (str, optional) String to match actor names against, e.g. "Cylinder.*"

Returns:
    list[str]: List containing all the names

Colosseum.simLoadLevel — Method

Loads a level specified by its name

Args: level_name (str) Name of the level to load

Returns: bool: True if the level was successfully loaded

Colosseum.simPause — Method

Pauses simulation

Args: is_paused (bool) True to pause the simulation, false to release

Colosseum.simPrintLogMessage — Function

Prints the specified message in the simulator's window.

If messageparam is supplied, then it's printed next to the message and in that case if this API is called with same message value but different messageparam again then previous line is overwritten with new line (instead of API creating new line on display).

For example, simPrintLogMessage("Iteration: ", to_string(i)) keeps updating same line on display when API is called with different values of i. The valid values of severity parameter is 0 to 3 inclusive that corresponds to different colors.

Args: message (str) Message to be printed message_param (str, optional) Parameter to be printed next to the message severity (int, optional) Range 0-3, inclusive, corresponding to the severity of the message

Colosseum.simRunConsoleCommand — Method

Allows the client to execute a command in Unreal's native console, via an API. Affords access to the countless built-in commands such as "stat unit", "stat fps", "open [map]", adjust any config settings, etc. etc. Allows the user to create bespoke APIs very easily, by adding a custom event to the level blueprint, and then calling the console command "ce MyEventName [args]". No recompilation of AirSim needed!

Args: command ([string]) Desired Unreal Engine Console command to run

Returns: [bool]: Success

Colosseum.simSetCameraFov — Function

Control the field of view of a selected camera

Args: cameraname (str) Name of the camera to be controlled fovdegrees (float) Value of field of view in degrees vehicle_name (str, optional) Name of vehicle which the camera corresponds to external (bool, optional) Whether the camera is an External Camera

Colosseum.simSetCameraPose — Function

Control the pose of a selected camera

Args: cameraname (str) Name of the camera to be controlled pose (Pose) Pose representing the desired position and orientation of the camera vehiclename (str, optional) Name of vehicle which the camera corresponds to external (bool, optional) Whether the camera is an External Camera

Colosseum.simSetDetectionFilterRadius — Function

Set detection radius for all cameras

Args: cameraname (str) Name of the camera, for backwards compatibility, ID numbers such as 0,1,etc. can also be used imagetype (ImageType) Type of image required radiuscm (int) Radius in [cm] vehiclename (str, optional) Vehicle which the camera is associated with external (bool, optional) Whether the camera is an External Camera

Colosseum.simSetDistortionParam — Function

Set single camera distortion parameter

Args: cameraname (str) Name of the camera, for backwards compatibility, ID numbers such as 0,1,etc. can also be used paramname (str) Name of distortion parameter value (float) Value of distortion parameter vehicle_name (str, optional) Vehicle which the camera is associated with external (bool, optional) Whether the camera is an External Camera

Colosseum.simSetDistortionParams — Function

Set camera distortion parameters

Args: cameraname (str) Name of the camera, for backwards compatibility, ID numbers such as 0,1,etc. can also be used distortionparams (dict) Dictionary of distortion param names and corresponding values {"K1": 0.0, "K2": 0.0, "K3": 0.0, "P1": 0.0, "P2": 0.0} vehicle_name (str, optional) Vehicle which the camera is associated with external (bool, optional) Whether the camera is an External Camera

Colosseum.simSetLightIntensity — Method

Change intensity of named light

Args: light_name (String) Name of light to change intensity (Real) New intensity value

Returns: bool: True if successful, otherwise false

Colosseum.simSetObjectMaterial — Function

Runtime Swap Texture API See https://microsoft.github.io/AirSim/retexturing/ for details Args: objectname (str) name of object to set material for materialname (str) name of material to set for object component_id (int, optional) : index of material elements

Returns: bool: True if material was set

Colosseum.simSetObjectMaterialFromTexture — Function

Runtime Swap Texture API See https://microsoft.github.io/AirSim/retexturing/ for details Args: objectname (str) name of object to set material for texturepath (str) path to texture to set for object component_id (int, optional) : index of material elements

Returns: bool: True if material was set

Colosseum.simSetObjectPose — Function

Set the pose of the object(actor) in the environment

The specified actor must have Mobility set to movable, otherwise there will be unend

functionined behaviour. See https://www.unrealengine.com/en-US/blog/moving-physical-objects for details on how to set Mobility and the effect of Teleport parameter

Args:
    object_name (str) Name of the object(actor) to move
    pose (Pose) Desired Pose of the object
    teleport (bool, optional) Whether to move the object immediately without affecting their velocity

Returns:
    bool: If the move was successful

Colosseum.simSetObjectScale — Method

Sets scale of an object in the world

Args: objectname (str) Object to set the scale of scalevector (airsim.Vector3r) Desired scale of object

Returns: bool: True if scale change was successful

Colosseum.simSetSegmentationObjectID — Function

Set segmentation ID for specific objects

See https://microsoft.github.io/AirSim/image_apis/#segmentation for details

Args: meshname (str) Name of the mesh to set the ID of (supports regex) objectid (int) Object ID to be set, range 0-255

                    RBG values for IDs can be seen at https://microsoft.github.io/AirSim/seg_rgbs.txt
is_name_regex (bool, optional) Whether the mesh name is a regex

Returns: bool: If the mesh was found

Colosseum.simSetTimeOfDay — Function

Control the position of Sun in the environment

Sun's position is computed using the coordinates specified in OriginGeopoint in settings for the date-time specified in the argument, else if the string is empty, current date & time is used

Args: isenabled (bool) True to enable time-of-day effect, false to reset the position to original startdatetime (str, optional) Date & Time in %Y-%m-%d %H:%M:%S format, e.g. 2018-02-12 15:20:00 isstartdatetimedst (bool, optional) True to adjust for Daylight Savings Time celestialclockspeed (float, optional) Run celestial clock faster || slower than simulation clock E.g. Value 100 means for every 1 second of simulation clock, Sun's position is advanced by 100 seconds so Sun will move in sky much faster updateintervalsecs (float, optional) Interval to update the Sun's position movesun (bool, optional) Whether || not to move the Sun

Colosseum.simSetTraceLine — Function

Modify the color and thickness of the line when Tracing is enabled

Tracing can be enabled by pressing T in the Editor || setting EnableTrace to True in the Vehicle Settings

Args: colorrgba (list) desired RGBA values from 0.0 to 1.0 thickness (float, optional) Thickness of the line vehiclename (string, optional) Name of the vehicle to set Trace line values for

Colosseum.simSetVehiclePose — Function

Set the pose of the vehicle

If you don't want to change position (or orientation) then just set components of position (or orientation) to floating point nan values

Args: pose (Pose) Desired Pose pf the vehicle ignorecollision (bool) Whether to ignore any collision || not vehiclename (str, optional) Name of the vehicle to move

Colosseum.simSetWeatherParameter — Method

Enable various weather effects

Args: param (WeatherParameter) Weather effect to be enabled val (float) Intensity of the effect, Range 0-1

Colosseum.simSpawnObject — Function

Spawned selected object in the world

Args: objectname (str) Desired name of new object assetname (str) Name of asset(mesh) in the project database pose (airsim.Pose) Desired pose of object scale (airsim.Vector3r) Desired scale of object physicsenabled (bool, optional) Whether to enable physics for the object isblueprint (bool, optional) Whether to spawn a blueprint || an actor

Returns: str: Name of spawned object, in case it had to be modified

Colosseum.simSwapTextures — Function

Runtime Swap Texture API

See https://microsoft.github.io/AirSim/retexturing/ for details

Args: tags (str) string of "," || ", " delimited tags to identify on which actors to perform the swap tex_id (int, optional) indexes the array of textures assigned to each actor undergoing a swap

                        If out-of-bounds for some object's texture set, it will be taken modulo the number of textures that were available
component_id (int, optional)
material_id (int, optional)

Returns: list[str]: List of objects which matched the provided tags and had the texture swap perfomed

Colosseum.simTestLineOfSightBetweenPoints — Method

Returns whether the target point is visible from the perspective of the source point

Args: point1 (GeoPoint) source point point2 (GeoPoint) target point

Returns: [bool]: Success

Colosseum.simTestLineOfSightToPoint — Function

Returns whether the target point is visible from the perspective of the inputted vehicle

Args: point (GeoPoint) target point vehicle_name (str, optional) Name of vehicle

Returns: [bool]: Success